+++ ACG/Scenegraph/PolyLineNodeT.hh
@@ -121,7 +121,7 @@ public:
   unsigned int availableDrawModes() const;
 
   /// update bounding box
-  void boundingBox(Vec3f& _bbMin, Vec3f& _bbMax);
+  void boundingBox(Vec3d& _bbMin, Vec3d& _bbMax);
 
   /// draw lines and normals
   void draw(GLState& /*_state*/, unsigned int _drawMode);
