+++ ACG/Scenegraph/TextNode.cc
@@ -83,7 +83,7 @@ TextNode::
 
 void
 TextNode::
-boundingBox(Vec3f& /*_bbMin*/, Vec3f& /*_bbMax*/)
+boundingBox(Vec3d& /*_bbMin*/, Vec3d& /*_bbMax*/)
 {
 }
